Funpay-Telegram-Gifts/requirements.txt
pyrogram
FunPayAPI
python-dotenv
colorlog